+++ Changes.md
@@ -10,6 +10,7 @@ Features
   - Improved comparison abilities, including the added ability to compare two render passes.
   - Added subsections for options and attributes, splitting them by category.
   - Added column resizing, including automatic resizing to show long attribute names.
+  - Added inspection of all shaders and parameters in shader networks.
 
 Improvements
 ------------